Recall table programs tested on Linux Mint. Fix: Correct path to activation script on Linux. Fix: Path to script directory is absolute.

+++ src/index/chm_tools/RecallTable.cpp
@@ -120,7 +120,7 @@ namespace chm {
 		s << '.';
 		print(convert<chr::microseconds>(elapsedCopy), s, 3);
 		s << '.';
-		print(elapsedCopy.count(), s, 3);
+		print((long long)(elapsedCopy.count()), s, 3);
 
 		s.copyfmt(streamState);
 	}
